pyproject.toml
[project]
name = "hotmart-ai-support"
version = "0.1.0"
description = "avaliação de Data Scientist da Hotmart"
authors = [
    {name = "evellynnicole",email = "mevellyn4@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "langchain-qdrant (==0.2.0)",
    "qdrant-client (==1.14.3)",
    "langchain-openai (==0.3.28)",
    "python-dotenv (==1.1.1)",
    "langchain (==0.3.26)",
    "langchain-core (==0.3.69)",
    "pandas (==2.3.1)",
    "fastembed (==0.7.1)",
    "langgraph (==0.5.3)",
    "fastapi (==0.116.1)",
    "uvicorn (==0.35.0)",
    "langchain-community (==0.3.27)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-cov = "^6.2.1"
taskipy = "^1.14.1"
ruff = "^0.12.3"

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'